+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/KubernetesClient.java
@@ -64,6 +64,7 @@ import io.fabric8.kubernetes.client.dsl.AuthorizationAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.AutoscalingAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.BatchAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.CertificatesAPIGroupDSL;
+import io.fabric8.kubernetes.client.dsl.ConfigMapResource;
 import io.fabric8.kubernetes.client.dsl.DiscoveryAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.EventingAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.ExtensionsAPIGroupDSL;
@@ -450,7 +451,7 @@ public interface KubernetesClient extends Client {
    *
    * @return MixedOperation object for ConfigMap related operations.
    */
-  MixedOperation<ConfigMap, ConfigMapList, Resource<ConfigMap>> configMaps();
+  MixedOperation<ConfigMap, ConfigMapList, ConfigMapResource> configMaps();
 
   /**
    * API entrypoint for LimitRange related operations. LimitRange (core/v1)